Reorganized TunnelCommunity class structure
