+++ Tribler/Core/Config/config.spec
@@ -20,6 +20,7 @@ enabled = boolean(default=True)
 
 [tunnel_community]
 enabled = boolean(default=True)
+hidden_seeding_enabled = boolean(default=True)
 socks5_listen_ports = string_list(default=list('-1', '-1', '-1', '-1', '-1'))
 exitnode_enabled = boolean(default=False)
 
@@ -29,7 +30,6 @@ matchmaker = boolean(default=True)
 ec_keypair_filename = string(default='')
 
 [trustchain]
-enabled = boolean(default=True)
 ec_keypair_filename = string(default='')
 live_edges_enabled = boolean(default=True)
