+++ Tribler/Test/Community/Market/Integration/test_market_base.py
@@ -99,7 +99,6 @@ class TestMarketBase(TestAsServer):
         self.config.set_dispersy_enabled(True)
         self.config.set_libtorrent_enabled(False)
         self.config.set_video_server_enabled(False)
-        self.config.set_trustchain_enabled(False)
         self.config.set_tunnel_community_enabled(False)
         self.config.set_market_community_enabled(False)
