+++ Tribler/Test/Core/Modules/RestApi/test_debug_endpoint.py
@@ -1,5 +1,3 @@
-import logging
-import logging.config
 import os
 
 from twisted.internet.defer import inlineCallbacks
@@ -8,7 +6,7 @@ import Tribler.Core.Utilities.json_util as json
 from Tribler.Test.Core.Modules.RestApi.base_api_test import AbstractApiTest
 from Tribler.Test.Core.base_test import MockObject
 from Tribler.Test.twisted_thread import deferred
-from Tribler.community.tunnel.hidden_community import HiddenTunnelCommunity
+from Tribler.community.hiddentunnel.hidden_community import HiddenTunnelCommunity
 from Tribler.dispersy.dispersy import Dispersy
 from Tribler.dispersy.endpoint import ManualEnpoint
 from Tribler.dispersy.member import DummyMember
