+++ Tribler/community/hiddentunnel/__init__.py
@@ -1,0 +1,4 @@
+"""
+This package contains code for the hidden tunnel community and is an extension built on top of the TunnelCommunity.
+It provides a way for users to seed anonymously.
+"""
